HR agent/requirements.txt
streamlit==1.37.1
openai==1.60.2
python-dotenv==1.0.1
pydantic==2.11.7